backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
